refactor: moved generators into own namespace
