+++ src/Generators/CategoriesJsonGenerator.php
@@ -1,13 +1,12 @@
 <?php
 
-namespace Photogabble\ConfusableHomoglyphs\Categories;
+namespace Photogabble\ConfusableHomoglyphs\Generators;
 
 use DateTime;
 use DateTimeZone;
 use Exception;
-use Photogabble\ConfusableHomoglyphs\Generator;
 
-class JsonGenerator implements Generator
+class CategoriesJsonGenerator implements Generator
 {
 
     /**
